Alright! It is resolved posts are successfully posted. but they are still not showing on the posts page.
You seem to have merged the Home and posts page which are supposed to be separate. leave the homepage as placeholder for now. As for the posts page. It is not displaying the posts that I just successfully posted. The create button is also not supposed to be on the sidebar, but on the posts page.